Split test_backend_tensorflow and create test_hypersphere_tensorflow
